requirements.txt
langchain
openai
pgvector
psycopg2-binary
sqlalchemy
python-dotenv
fastapi
uvicorn
streamlit